Add optional database file consistency checking when compacting.
This code will scan a live or a compacted database file, reading
all by_id and by_seq btree nodes and document objects, checking them
for correct structure and byte level consistency. If anything is found
to be incorrect, it exits the Erlang process for debugging purposes.
This is not intended for production use, but to check Couchstore
writes and compactions for correctness in near real-time.

Change-Id: Iad10aac6a9ea95de3e467f1f867881fdbb8d9a54